Combine appointments and schedule microservices

+++ appointments_microservice/microservice/repository/repository.py
@@ -4,6 +4,13 @@ from domain_objects import Appointment, Message, Reason, OptMessage
 from domain import DomainLayer
 from uuid import UUID
 import psycopg2
+import time
+from cassandra.cluster import Cluster
+from cassandra.cluster import NoHostAvailable
+from cassandra.query import ordered_dict_factory
+import requests
+import pandas as pd
+import json
 
 
 
@@ -12,8 +19,50 @@ class RepositoryLayer:
         self.connection = psycopg2.connect(database='test_db', user='postgres', 
                         password='postgres', host='postgres-1')
         self.cursor = self.connection.cursor()
+    #     self.session = None # for Cassandra
+
+    # def execute(self, query, row):
+    #     prepared = self.session.prepare(query)
+    #     self.session.execute(prepared, row)
+
+    def delete_timeslot(self, appointment: Appointment):
+        app_doctor_id = appointment.doctor
+        app_date = appointment.date.strftime('%Y-%m-%d %H:%M:%S')
+        schedule_url = "http://scheduleService:8080/get_timeslots"
+        get_json = {"doctor": str(app_doctor_id), "date": str(app_date)}
+        try:
+            response = requests.get(url=schedule_url, json=get_json).json()
+            response_data = json.loads(response.strip())
+            response_2 = json.loads(response_data)
+            # print("get", response_2)
+            # print(type(response_2))
+            # print(response_2[0])
+            # print(response_2[0]["timeslot_id"])
+        except Exception as e:
+            print(f"Failed to get from Timeslots: {str(e)}")
+            return False
+        
+        if len(response_2) != 1:
+            print(f"There should be one timeslot for such appointment, insted {len(response)} were found")
+            return False
+        timeslot_id = response_2[0]["timeslot_id"]
+        # print(timeslot_id)
+        schedule_url = "http://scheduleService:8080/delete_timeslot"
+        post_json = {"timeslot_id": str(timeslot_id)}
+        try:
+            response = requests.post(url=schedule_url, json=post_json)
+            print("post", response.json())
+        except Exception as e:
+            print(f"Failed to post to Timeslots: {str(e)}")
+            return False
+        return True
+
 
     def save_appointment(self, appointment: Appointment):
+        if not self.delete_timeslot(appointment):
+            print("Could not remove timeslot for the appointment")
+            return False
+
         # save appointment to rdbms
         sql = "INSERT INTO FutureAppointments (appointment_id, doctor_id, patient_id, type, date) VALUES (%s, %s, %s, %s, %s);"
         try:
@@ -104,9 +153,34 @@ class RepositoryLayer:
         if result == []:
             return False
         return True
+    
+
+    def return_timeslot(self, appointment_id: UUID):
+        sql = f"SELECT * FROM FutureAppointments WHERE appointment_id = '{appointment_id}'"
+        self.cursor.execute(sql)
+        result = self.cursor.fetchall()
+        if result:
+            appointment = result[0]
+            app_doctor_id = appointment[1]
+            app_start_date = appointment[4]
+            app_end_date = app_start_date + pd.Timedelta(minutes=15)
+            schedule_url = "http://scheduleService:8080/new_timeslots"
+            post_json = {"doctor": app_doctor_id, "start_date": str(app_start_date), "end_date": str(app_end_date)}
+            try:
+                response = requests.post(url=schedule_url, json=post_json)
+                print(response.json())
+            except Exception as e:
+                print(f"Failed to post to Timeslots: {str(e)}")
+                return False
+            return True
+        else:
+            print("No such appointment")
+        return False
         
     
     def delete_future_appointment(self, appointment_id: UUID, commit=True):
+        if not self.return_timeslot(appointment_id):
+            return False
         # delete from rdbms
         try:
             self.cursor.execute("DELETE FROM FutureAppointments WHERE appointment_id = %s", (appointment_id,))
@@ -135,6 +209,10 @@ class RepositoryLayer:
         print(appointment)
         appointment = appointment[0]
         status = self.delete_future_appointment(appointment_id, commit=False)
+
+        if not self.delete_timeslot(appointment):
+            print("Could not remove timeslot for the appointment")
+            return False
         if status:
             sql = "INSERT INTO PastAppointments (appointment_id, doctor_id, patient_id, type, date) VALUES (%s, %s, %s, %s, %s);"
             try:
@@ -144,3 +222,27 @@ class RepositoryLayer:
             except:
                 return False
         return False
+    
+
+    def connect_cassandra(self, keyspace=None, max_retries=100, retry_delay=5):
+        host='cassandra-node'
+        port=9042
+        retry_count = 0
+        while retry_count < max_retries and not self.session:
+            try:
+                cluster = Cluster([host], port=port)
+                if keyspace is None:
+                    session = cluster.connect()
+                else:
+                    session = cluster.connect(keyspace, wait_for_all_pools=True)
+                    session.row_factory = ordered_dict_factory
+                    session.execute('USE schedule')
+                print("Successful connection!")
+            except Exception as e:
+                print(f"Failed to connect to Cassandra: {str(e)}")
+                print("Retrying connection...")
+                retry_count += 1
+                time.sleep(retry_delay)
+
+        if retry_count == max_retries:
+            print("Failed to connect to Cassandra after multiple retries. Exiting...")
